+++ src/DeviceDistance.h
@@ -5,7 +5,7 @@
 
 class DeviceDistance : public Device {
   public:
-    DeviceDistance(uint32_t device_id, uint8_t write_pin, uint8_t read_pin);
+    DeviceDistance(uint32_t device_id, uint8_t write_pin, uint8_t read_pin, uint32_t poll);
     bool is_output();
     void loop();
     void action(JsonObject doc);